Pull out MockJWTDecodeWithRole
