+++ api/app/tests/hfi/test_hfi_admin.py
@@ -28,6 +28,7 @@ add_stations_json = {
     "removed": []
 }
 
+
 post_admin_stations_url = '/api/hfi-calc/admin/stations'
