+++ api/app/tests/utils/mock_jwt_decode_role.py
@@ -1,0 +1,24 @@
+class MockJWTDecodeWithRole:
+    """ Mock pyjwt module with role """
+
+    def __init__(self, role):
+        self.decoded_token = {
+            "preferred_username": "test_username",
+            "resource_access": {
+                "wps-web": {
+                    "roles": [
+                        role
+                    ]
+                }
+            }}
+
+    def __getitem__(self, key):
+        return self.decoded_token[key]
+
+    def get(self, key, _):
+        "Returns the mock decoded token"
+        return self.decoded_token[key]
+
+    def decode(self):
+        "Returns the mock decoded token"
+        return self.decoded_token
